fix: avoid "instanceof ConstantEvaluateError"
I'm not entirely sure of the problem here, but in the released version
of groq-js 1.0.0 in Node.js 18 I'm seeing a strange error where it
appears that "err instanceof ConstantEvaluateError" returns false.
This is a bit hard to test for since it seems to be caused by the
way Rollup builds the file.

(cherry picked from commit 5a5e649ebcdfed2e8979469816b564a200300511)