feat(parser): Add geo namespace with noop
The geo function is part of the spec as an extension, but is not
included as part of groq-js. This makes groq-js non-compliant in
accordance with the spec with extensions. This is especially problematic
in places where groq-js is relied upon to safely parse queries but not
evaluate anything, eg. sanity codegen.

As functions in the text namespace are already implemented with noops to
not break the parser, I have done the same for all functions in the geo
namespace.